refactor: split commandsprovider and webview provider
