+++ package.json
@@ -345,7 +345,7 @@
         "eslint": "^8.37.0",
         "eslint-config-prettier": "^8.8.0",
         "husky": "^8.0.3",
-        "lint-staged": "^13.2.0",
+        "lint-staged": "^13.3.0",
         "lockfile-lint": "^4.10.1",
         "prettier": "^2.8.7",
         "typescript": "^5.0.3"
