+++ src/extension.ts
@@ -7,9 +7,13 @@ export function activate(context: vscode.ExtensionContext) {
   const dataFileEventEmitter = new vscode.EventEmitter<Array<DataFile>>();
 
   let dataFileWatcher = new DataFileWatcher(getConfig().get('basePath'), dataFileEventEmitter);
+
+  const monthSummaryProvider = new provider.MonthSummaryProvider(dataFileEventEmitter.event);
   context.subscriptions.push(
     ...[
-      new provider.CommandsProvider(dataFileEventEmitter.event),
+      new provider.CommandsProvider(),
+      monthSummaryProvider,
+      new provider.DayEditorProvider(monthSummaryProvider),
       new provider.DataFileTreeProvider(dataFileEventEmitter.event),
       new provider.StatusBarItemProvider(dataFileEventEmitter.event),
       new provider.ReminderProvider(dataFileEventEmitter.event),
