+++ src/provider/index.ts
@@ -1,6 +1,8 @@
 export * from './actions';
 export * from './commandsProvider';
 export * from './dataFileTreeProvider';
+export * from './dayEditorProvider';
+export * from './monthSummaryProvider';
 export * from './reminderProvider';
 export * from './statusBarItemProvider';
 export * from './treeItem';
