+++ tsconfig.json
@@ -1,8 +1,10 @@
 {
 	"compilerOptions": {
 		"module": "commonjs",
-		"target": "es2019",
-		"lib": ["ES2019"],
+		"target": "ES2022",
+		"lib": [
+			"ES2022"
+		],
 		"outDir": "dist",
 		"esModuleInterop": true,
 		"sourceMap": true,
